add confirmation alert to unregister

+++ projects/mainsite/app/scripts/controllers/dashboard.js
@@ -97,16 +97,19 @@ angular.module('shaastra2016App')
       };
 
       $scope.unregisterEvent = function (team, event, eventIndex, teamIndex) {
-        $scope.eventUnRegisterMessage = ' -- Working...';
-        $http.delete('http://localhost:8001/api/registrations/' + team._id + '/' + event._id)
-          .then(function (response) {
-            if(response.status === 204) {
-              $scope.eventUnRegisterMessage = '';
-              $scope.all_teams[teamIndex].eventsRegistered.splice(eventIndex, 1);
-            } else {
-              $scope.eventUnRegisterMessage = 'Some error occurred';
-            }
-          });
+        var result = confirm("Are you sure you want to Unregister?");
+        if(result) {
+          $scope.eventUnRegisterMessage = ' -- Working...';
+          $http.delete('http://localhost:8001/api/registrations/' + team._id + '/' + event._id)
+            .then(function (response) {
+              if(response.status === 204) {
+                $scope.eventUnRegisterMessage = '';
+                $scope.all_teams[teamIndex].eventsRegistered.splice(eventIndex, 1);
+              } else {
+                $scope.eventUnRegisterMessage = 'Some error occurred';
+              }
+            });
+        }
       };
 
       $scope.showTeamRequire = function() {
@@ -189,32 +192,38 @@ angular.module('shaastra2016App')
 		};
 
 	$scope.leaveTeam = function (index) {
-		$scope.teamBlockMessage = ' -- Working...';
-		var teamId = $scope.all_teams[index]._id;
-		$http.post('http://localhost:8001/api/teams/leave/'+ teamId)
-		  .then(function (response) {
-  			console.log(response);
-  			if(response.status === 200) {
-  				$scope.teamBlockMessage = '';
-  				$scope.all_teams.splice(index, 1);
-  			} else {
-  				$scope.teamBlockMessage = 'Some error occurred!';
-  			}
-  		});
+    var result = confirm("Are you sure you want to Leave Team?");
+    if(result) {
+  		$scope.teamBlockMessage = ' -- Working...';
+  		var teamId = $scope.all_teams[index]._id;
+  		$http.post('http://localhost:8001/api/teams/leave/'+ teamId)
+  		  .then(function (response) {
+    			console.log(response);
+    			if(response.status === 200) {
+    				$scope.teamBlockMessage = '';
+    				$scope.all_teams.splice(index, 1);
+    			} else {
+    				$scope.teamBlockMessage = 'Some error occurred!';
+    			}
+    		});
+    }
 	};
 
   $scope.deleteTeam = function (index) {
-    $scope.teamBlockMessage = ' -- Working...';
-    var teamId = $scope.all_teams[index]._id;
-    $http.delete('http://localhost:8001/api/teams/' + teamId)
-      .then(function (response) {
-        if(response.status === 204) {
-          $scope.teamBlockMessage = '';
-          $scope.all_teams.splice(index, 1);
-        } else {
-          $scope.teamBlockMessage = 'Some error occurred';
-        }
-      });
+    var result = confirm("Are you sure you want to Delete Team?");
+    if(result) {
+      $scope.teamBlockMessage = ' -- Working...';
+      var teamId = $scope.all_teams[index]._id;
+      $http.delete('http://localhost:8001/api/teams/' + teamId)
+        .then(function (response) {
+          if(response.status === 204) {
+            $scope.teamBlockMessage = '';
+            $scope.all_teams.splice(index, 1);
+          } else {
+            $scope.teamBlockMessage = 'Some error occurred';
+          }
+        });
+    }
   };
 
 	$scope.lists=[{
